refactor: Separated Timeline and Weekday components into specific class files
